Reorganization of AST
* Break "Definitions" and "Actions" out of AST as traits
* Attempt to fix scala formatting
* Prune .gitignore
