+++ .gitignore
@@ -11,7 +11,5 @@ target
 **/*.gzip
 /chroma-riddl-lexer/go.sum
 /.riddl-timestamp
-/foo
 doc/src/main/hugo/public/
 *.lock
-examples/src/riddl/ReactiveBBQ/static/images/logo.png
