+++ .scalafmt.conf
@@ -1,61 +1,22 @@
 version=3.6.0
 preset = default
 maxColumn = 80
-
-indent.defnSite = 2
 docstrings.style = SpaceAsterisk
 runner.dialect = scala213source3
-
-// If true, the margin character | is aligned with the opening
-// triple quote in interpolated and raw string literals.
-assumeStandardLibraryStripMargin = true
-includeCurlyBraceInSelectChains = true
-includeNoParensInSelectChains = true
-
-danglingParentheses.preset = true
+indent.main = 2
 
 optIn {
-  breakChainOnFirstMethodDot = false
+  breakChainOnFirstMethodDot = true
   annotationNewlines = false
   configStyleArguments = true
 }
 
-
-continuationIndent {
-  callSite = 2
-  defnSite = 2
-  ctorSite = 2
-}
-
-// Set to `none` setting to minimize git diffs/conflicts from
-// renamings and other refactorings.
-align = none
 align {
-  stripMargin = true
-  tokens = [caseArrow]
-  arrowEnumeratorGenerator = false
-  openParenCallSite = false
   openParenDefnSite = false
 }
 
-
-newlines {
-  source = fold
-  alwaysBeforeElseAfterCurlyIf = false
-}
-
-rewrite {
-  rules = [ExpandImportSelectors, PreferCurlyFors]
- redundantBraces.stringInterpolation = true
-}
-
-verticalMultiline {
-  arityThreshold = 100
-  newlineAfterOpenParen = false
-  atDefnSite = true
-  newlineAfterOpenParen = false
-}
-
+indent.ctorSite = 2
+indent.defnSite = 2
 
 // Configure which source files should be formatted in this project.
 
@@ -68,11 +29,3 @@ project {
   excludeFilters = []
 }
 
-// Miscellaneous
-
-binPack.literalArgumentLists = false
-
-
-spaces {
-  inImportCurlyBraces = true
-}
