+++ language/src/main/scala/com/reactific/riddl/language/ast/Expressions.scala
@@ -6,6 +6,8 @@
 
 package com.reactific.riddl.language.ast
 
+import com.reactific.riddl.language.parsing.Terminals.*
+
 import scala.collection.immutable.ListMap
 
 /** A trait for inclusion into AST object in AST.scala */
