+++ language/src/main/scala/com/reactific/riddl/language/parsing/ContextParser.scala
@@ -7,6 +7,7 @@
 package com.reactific.riddl.language.parsing
 
 import com.reactific.riddl.language.AST.*
+import Terminals.*
 import fastparse.*
 import fastparse.ScalaWhitespace.*
