+++ language/src/main/scala/com/reactific/riddl/language/parsing/FunctionParser.scala
@@ -9,6 +9,7 @@ package com.reactific.riddl.language.parsing
 import com.reactific.riddl.language.AST.*
 import fastparse.*
 import fastparse.ScalaWhitespace.*
+import Terminals.*
 
 /** Unit Tests For FunctionParser */
 private[parsing] trait FunctionParser extends CommonParser with TypeParser with GherkinParser {
