+++ language/src/main/scala/com/reactific/riddl/language/parsing/GherkinParser.scala
@@ -9,6 +9,7 @@ package com.reactific.riddl.language.parsing
 import com.reactific.riddl.language.AST.*
 import fastparse.*
 import fastparse.ScalaWhitespace.*
+import Terminals.*
 
 /** Parsing rules for feature definitions This is based on Cucumber's Gherkin
   * language.
